Update patch set 2
Patch Set 2:

(1 comment)

The community can not create properties that are limited to geo shapes, if there is no data type for that, just a generic "whatever is in this generic namespace". Note that this *is* different from the file namespace. All file types in the file namespace are *presentational*, while both .tab and .map are raw data. But *very* different raw data that can not be used by the same consumers. Basically the question is: Will there ever be a consumer that is able to consume both .tab and .map data the same time? I don't think so.

Discussed with Lydia today. She wants to talk to Yuri and better understand what the plans with this namespace are.

Patch-set: 2
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=0