Update patch set 2
Patch Set 2:

(4 comments)

Intention of structuring interface this way is to put constrains on developer so he/she to write more modular code and also to be able to know structure of request in `createChangeOp. Because we cannot define top level structure as soon as it can be extended by extensions, but we can define structure on top key level.

Daniel had a concern that if we restrict interface that way, then if we ever need a ChangeOp that needs to access two or more top level fields, then we won't be able to do it, but we failed to come up with an example where it can be possibly needed. (If anyone has an example of that case please share)

I gave it more thought and can say that it is not true.
If one will want to access data in two keys he can create a deserializer for each key which will contain reference to some builder and push required data there. As soon as builder got all it needs it can create needed ChangeOp.

Solution is a bit complicated, but the purpose itself is not what I would call desirable.

The only real constraint of this interface that I currently found is that implementation won't be able to handle data for unknown keys, but anyway it will/might have unknown structure and can not imagine how it can be properly handled.

Patch-set: 2