Update patch set 5
Patch Set 5: Code-Review+1

I just realized that the test was using the same global state before, so this issue was not introduced with this patch. No +2 because of the renamed variable. It was a long time ago that I tried to apply a naming scheme. These should all be called "$wikibaseRepo" for consistency, because "$repo" alone can mean different things (typically the string identifier of a data repository).

Patch-set: 5
Label: Code-Review=+1