Update patch set 2
Patch Set 2: Code-Review-2

I believe this patch is removing to much helpful knowledge from the comments I wrote. For example, implementations should be aware that the provided array is *not* necesarrily minified and only represents actual changes. More often it will be a *full* entity with changed elements as well as all unchanged elements.

The comment about the "validity" of the provided array can be rephrased, but must stay. For example, if a label deserializer starts processing the "label" sub-structure, but finds it is incomplete, it should throw an exception. I'm not sure if this is the right one, though.

I like the @see tags replacing the class-level comment. Please submit this as a separate patch.

I'm absolutely happy with a NullChangeOp, which means we can remove the ugly "@return …|null" I introduced. Please go ahead and create a separate patch for this.

However, I would vote against singleton. It's an entirely stateless object. What getInstance does probably wastes more ressources than a straight instantiation with the standard constructor. I find a straight "new …" more readable. When I find a "getInstance" it's always a hint that something expensive is going on, which is not the case here.

Finally, I'm still a bit torn but vote against two methods (may it be an "isDeserializerFor" or the "getKey" in this patch). I discussed this with Daniel yesterday and realized I did not considered an argument: It should be possible to create an ItemChangeOpDeserializer, a PropertyChangeOpDeserializer, and so on. Each would obviously use the existing deserializers for labels, statements and so on, but this would be an implementation detail of an entity-type agnostic deserializer. The proposal in this patch would not only remove this freedom, it would also create the situation where different extensions will register the same array keys. How do two extensions know which array keys the other already covers? They can not do anything about this, but will create identical ChangeOps, which will conflict and surface as edit conflicts.

Patch-set: 2
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=-2