Update patch set 2
Patch Set 2:

Thiemo, thanks for the comments.

 > Finally, I'm still a bit torn but vote against two methods (may it be an "isDeserializerFor" or the "getKey" in this patch). I discussed this with Daniel yesterday and realized I did not considered an argument: It should be possible to create an ItemChangeOpDeserializer, a PropertyChangeOpDeserializer, and so on. Each would obviously use the existing deserializers for labels, statements and so on, but this would be an implementation detail of an entity-type agnostic deserializer. The proposal in this patch would not only remove this freedom, it would also create the situation where different extensions will register the same array keys. How do two extensions know which array keys the other already covers? They can not do anything about this, but will create identical ChangeOps, which will conflict and surface as edit conflicts.

So when I was thinking do we want have a ChangeOpDeserializer that needs to access multiple keys something like ItemChangeOpDeserializer was the first thing that came to my mind.
It is clear to me such service should not be accessing 'label', 'alias' etc keys directly - it would rather use LabelChangeOpDeserializer, AliasChangeOpDeserializer and so on. So for the case of those former services the new interface would be fit well.

But I think Thiemo has a point when it comes to the broader picture. We want to have EditEntity class get a pointer from entity type definition to what ChangeOpDeserializer to use in order to edit the entity. It seems to me it would be much cleaner if EditEntity expected a single implementation of ChangeOpDeserializer and e.g. item type defintion would return ItemChangeOpDeserializer to it, than EditEntity expecting string-indexed array of ChangeOpDeserializers (like [ 'label' => LabelChangeOp, 'alias' => AliasChangeOp, ... ], and than looping through the input array, and passing particular parts to it to deserializers relevant for the particular key, and then somehow combining all ChangeOp instance it got from multiple deserializers. At least now when I think of, the latter seems quite nasty to me.

It is a shame we don't have even a draft code making use of ChangeOpDeserializer, it might have been easier to see how the new proposed interface applies.

So I am with Thiemo, I think we should consider staying with the existing interface.

Patch-set: 2
Reviewer: WMDE-leszek <2626@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=0