Remove rare @group tags from tests
Most of these are only used a single time. This does not add any value,
when you can as well select the file by it's name.

Some have actual typos.

"Claim" is a word we are not using any more, because most of this code
is not dealing with claims but with statements.

Some are obscure and inconsistently set anyway.

Change-Id: I43bbc64206606c9ac1e65815b3e252bd5d101fdd