Remove all @since tags
As soon as we don't have releases and versioning for quiet some time
and sometimes in patches breaking changes being made, but version is not
updated.
Also there are only 0.1.x and 0.2.x branches but not for other versions.
And finally, one can use git history to find out when change
was introduced.

Tags were removed with following pattern: `^\s*\*\s*@since\s+0\.\d\s*\n`

Change-Id: I07bbc631714e9b52726c86ef74149565a2ab7c52