Update patch set 2
Patch Set 2: Code-Review-2

Sorry Aleksey, but what motivated this patch? How does it *hurt* to still have this information? The fact that we have not updated the version number for a while can not be the only reason for deleting *all* of this.

Setting a temporary -2, because this clearly needs discussion, most preferably in person and not here on Gerrit.

Patch-set: 2
Label: Code-Review=-2