Update patch set 2
Patch Set 2:

> How does it *hurt* to still have this information? 
Can't say that it hurts a lot, but for last two weeks I ran into a confusion with @since around three times: 2 times as contributor and 1 time as a reviewer. Also less code (or documentation, or anything else that needs to be looked after) we have the easier to support it.

 > The fact that we have not updated the version number for a while can not be the only reason for deleting *all* of this.
If we want it to mean something then we have to start using versioning schema and *really* follow it, otherwise I can't imagine how this information can be useful.

If someone knows how this information is currently used, please, can you share?

Patch-set: 2