Update patch set 2
Patch Set 2:

I agree that @since tags are quite useless if we never increment the version number. And since we don't have releases, we never do.

I propose to start using the MediaWiki version. Since Wikibase has no releases, it's basically bound to MediaWiki releases. So let's use MediaWiki's versioning.

But let's please take this discussion to wikidata-tech (and/or phabricator). Gerrit is not the right place for this kind of decision.

Patch-set: 2