Update patch set 1
Patch Set 1: Code-Review+2

If a method is guaranteed to return either an object or null, it's very common to use "if ( !$var )" in our code base. This is not an error that needs fixing.

The string comparison is bad and the reason why I merge this.

Patch-set: 1
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+2