Much larger OOUI style close icon for inputextender
We said this is relevant because we want to remove the sliding animation,
and users now actually need this button for the first time.

The icon file is copy pasted for now, until this becomes a real OOUI
widget.

This patch also fixes the position of the close icon for RTL languages.

Change-Id: I3e2dc509b7d11b81a0bf39f9d9d5b1e0836be116