Update patch set 4
Patch Set 4:

This is change is neccessary for a hook which Echo listens to and sends notifications from. So the hook is the only place where we can assert that it's correct. But the hook is run from a deferred update and I'm not sure how tests deal with them. (It seems to me that they are just ignored.) Quite desperate...

Patch-set: 4
Reviewer: Matěj Suchánek <2664@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0