Update patch set 2
Patch Set 2:

I believe it is of not much value to create separate Phabricator tickets for refactorings that are part of an other task. You are right, this patch touches much more than the ListEntityUsage API module. Nevertheless, this patch is a direct result of T154900. As said, splitting is an option, but you have to tell me how.

Patch-set: 2
Label: Verified=0