Update patch set 17
Patch Set 17: Code-Review+1

(1 comment)

Looks good to me know. The new code in Api\EditEntity is not tested (which is not a surprise as there is no implementation of ChangeOpDeserializer around yet).
One thing that could still be improved is the documention of EntityTypeDefinitions::getChangeOpDeserializerCallbacks (see my comments; I am not sure if I was clear enough there - I'd fine with changing the doc myself if you prefer, just tell me).
I'd be fine with both merging this as it is now, or having this up until we have some ChangeOpDeserializer implementation and could test the new code in EditEntity class with e.g. MediaInfo entity type? The former might be preferred, as getting to the state where the latter would be possible might still take quite some time.

Patch-set: 17
Label: Code-Review=+1