Update patch set 3
Patch Set 3:

I agree that the task should probably not be linked in the commit message.
Having changes like this link to tickets that are only really tenuously linked adds confusion to phabricator, for example, looking at the ticket now we will see it has had a patch submitted (and soon merged). But when looking at the change it is actually only tenuously linked.

Patch-set: 3