Update patch set 19
Patch Set 19: Code-Review+1

I only reviewed PS19, and believe it is very nice and ready to be merged. Except for a very minor refactoring I uploaded as a separate patch, see Iad93777. No +2 yet because I want to give others a bit more time to have a look at this.

I also believe this is missing a test that actually runs the new code in EditEntity.

Patch-set: 19