Update patch set 12
Patch Set 12: Code-Review-2

(2 comments)

I've done some changes in PS12 but I still consider this a work-in-progress patch. There are few things listed as TODOs in the commit message (tests for the watcher part of the change; splitting the big patch set to separate patches: DataRetrievalServiceFactory interface, making DispatchingServiceFactory a Watcher, using DataRetrievalServiceFactory from client in repo; figuring out how to properly test the optional dispatching service usage in repo). And I am still not entirely sure if the approach here is the right one to take.

Patch-set: 12
Label: Code-Review=-2