Update patch set 1
Patch Set 1: Code-Review+1

(2 comments)

Looks fine, except I'm not happy with the somewhat "random" order of the constructor parameters. I understand that the order in PS1 just uses the existing order of the private fields in the class. I believe we can do better. We are changing the constructor anyway.

* $siteLookup is the most basic lookup for this special page.
* $siteLinkGroups is a parameter for $siteLinkTargetProvider and should follow that.
* $labelDescriptionLookupFactory is for the $badgeItems only and should be grouped together.
* $siteLinkChangeOpFactory is for the final edit.

I would suggest this order:

 $siteLookup
 $siteLinkTargetProvider
 $siteLinkGroups
 $badgeItems
 $labelDescriptionLookupFactory
 $siteLinkChangeOpFactory

I can upload a follow up that does this if you prefer. Please tell me.

Patch-set: 1
Label: Code-Review=+1