Update patch set 1
Patch Set 1: Code-Review-1

Generally fine, but this should be rebased on top of If7c7b14 and use the same code path I suggested there. Otherwise this method is accessing user input from a scope it should not use. This is hard to understand and can easily be refactored, see If7c7b14.

Patch-set: 1
Label: Code-Review=-1