Update patch set 1
Patch Set 1: Code-Review-1

For me they fail :)

There is similar problem in ApiJsonFormatTest and ApiXmlFormatTest, as there was in the patch related to Api\GetEntities.

ApiJsonFormatTest::testSetLabelJsonFormat (and similar for the XML format tests) try to get an instance of Api\SetLabel using the generic ApiFormatTestCase::getApiModule which assumes there are no dependecies injected to the constructor, which results in the fatal error.

Patch-set: 1
Label: Code-Review=-1