Update patch set 5
Patch Set 5:

I forgot to talk about the parameter order in the constructor. I can see you reused the order of the private properties. Which is fine. It's just that it does not seem to follow any logic. But there are so many and so different parameters here, it seems impossible to find the most "natural" order. I'm not going to even try. ;-)

Patch-set: 5