Update patch set 1
Patch Set 1:

I disagree, for a reason: This class is *so* unspecific, it will never be used anywhere else *because* of this. People are much more likely to re-create a new, more specific class if they ever need something like this in an other context. And I would encourage them to do so.

I do not think "Updates" is a good namespace name. What should it contain? What do you expect to find in such a namespace?

I could live with a "Deferred" namespace. This is where the interface is in core.

Patch-set: 1