Update patch set 1
Patch Set 1: Code-Review-1

Ah, I see.

The language is a property from the context. But this context is null on construction time. It is set with a setter *after* the special page is created.

The SpecialPage::getLanguage() call boils down to RequestContext::getMain(), and then boils down to just returning $wgContLang. (Not manually debugged, but this is how I read the code.) I believe you should use $wgContLang instead of $wgLang and this will be fine.

Patch-set: 1
Label: Code-Review=-1