Update patch set 1
Patch Set 1: Code-Review+1

"Deferred" seems misleading. There are deferred data updates, but not all of them are deferred. This adapter can be used with updates that are not deferred. In fact, data update classes should not know if they are executed immediately, deferred, or queued. That's up to the code that creates them, it's not a property of the class.

I'm fine with moving these to the Content namespace. DataUpdates are the mechanism that is used to store derived information to the database when the primary Content blob is changed. They are specific to a given content model, and are created and controlled by the ContentHandler for that content model.

+1 at a first glance. Will need to look closer for a +2, but i'm fine with someone else doing that.

Patch-set: 1
Label: Code-Review=+1