Update patch set 3
Patch Set 3: Code-Review+2

See I2a09062 for the test failure. I was able to reproduce it locally for a few minutes, then it disappeared. In case you want to dig into this, be aware: The stack trace shows the the error happens in BetaFeaturesHooks, and is triggered by Wikibase\Repo\Tests\Content\EntityContentFactoryTest. *But* this is not all. This happens because of what many, many other extensions do with the GetBetaFeaturePreferences hook.

Patch-set: 3
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+2