Account for entity change events in DispatchingServiceFactory and RepositoryServiceContainer
Some services, as for instance the prefetching
EntityMetaDataAccessor used by EntityRevisionLookup
in Repo's SqlStore can be watched by EntityWatchers
in order to update service's cache accordingly
when relevant entities are changed.

DispatchingServiceFactory should account for this.
DispatchingServiceFactory delegates the event to
the RepositoryServiceContainer configured for the
repository the change entity belongs to.
RepositoryServiceContainer in turn passes the event
to all configured services that implement
EntityStoreWatcher interface, so underlying
caches etc. could be updated accordingly.

Bug: T155633
Change-Id: Id5dfc9ee9ae36a43f63b10e3ab195c66e0265af8