Make DispatchingServiceFactory implement EntityStoreWatcher
Prefetching EntityMetaDataAccessor used by EntityRevisionLookup
can be watched by EntityWatchers, so that the cache
is updated accordingly when prefetched entities are changed.
For instance Repo's SqlStore instances do this.

DispatchingServiceFactory should also allow its users
to update prefetcher's cache when entity is updated.
In order to do that DispatchingServiceFactory implements
EntityStoreWatcher interface. When cache update request
comes to DispatchingServiceFactory it delegates the request
to the prefetching meta data accessor configured
for the particular repository.

Each repository's EntityRevisionLookup relies on
the prefetcher mentioned above, so whenever underlying
cache is purged, it is also reflected in results
EntityRevisionLookup returns.

Bug: T155633
Change-Id: Id5dfc9ee9ae36a43f63b10e3ab195c66e0265af8