Add factory of RepositoryServiceContainers
RepositoryServiceContainer creation logic is moved from
DispatchingServiceFactory to the dedicated
RepositoryServiceContainerFactory class.

Along with clear separation of concerns, DispatchingServiceFactory
no longer depends on WikibaseClient. RepositoryServiceContainerFactory
still is bound to WikibaseClient but the bound is slightly less
strong than it used to be in DispatchingServiceFactory.

Also using DispatchingServiceFactory in tests is now easier as
test double of RepositoryServiceContainerFactory could be used
to provide mock containers.

Change-Id: I733b118713620e662c5c0947bf2d3c98e06a859c