[DNM] Add tests for federation services and config
Note: those tests are created mostly for demonstration
purposes of work done on T149580, and should not be
checked in as they're now.

Currently tests mock database queries performed by actual
services reading from the database. This should be possibly
change to have actual entity creation/editing instead of
relying on mocked DB queries.

Structure of tests needs more thought (among other things,
code in Repo and Client tests should not be in 95% equal),
and test coverage should be also expanded.

Tests also cover the case of the optional use of services
configured in the Client component in WikibaseRepo.

Change-Id: Ia2361ae56ab5222c71f14f69a0f591e198363b97