Update patch set 15
Patch Set 15: Code-Review-1

(1 comment)

I've just noticed that this patch introduces dependency on Client to the Repo.
So now we have circular dependency between packages which can bring a lot of problems in the future, especially if we decide to split client and repo.
I guess an abstraction layer should be provided to reverse dependency direction.

Patch-set: 15
Label: Code-Review=-1