Update patch set 15
Patch Set 15:

(1 comment)

> I've just noticed that this patch introduces dependency on Client to the Repo.

Indeed. This is the point of this patch, and it intended. I've unintentionally removed the word "Client" from the first line of the commit message when doing last changes to the patch, so the intention might not have been obvious. I've restored it. The rest of commit message have been mentioning this, though. Still, maybe more on that should be said in the commit desc.

> So now we have circular dependency between packages which can bring a lot of problems in the future, especially if we decide to split client and repo.
> I guess an abstraction layer should be provided to reverse dependency direction.

It is problematic, yes. There have been some problems even when introducing this dependency (in this patch). On the other hand it seemed to be the best way we have to allow usage of "dispatching" services in Repo (thus make Repo able to access foreign entity data) without duplicating the code we've created in Client on Repo's side.

Generally we thought we should reconsider the Client vs Repo split. Theoretically they are separate but as we noticed, Client already depends a bit on Repo. Also, logically repo is kind of a client, so the division is not always clear.
OTOH, there is (quite some) code under Client namespace which is specific for client-wiki use, so it does not make sense to always depend on all of the client code in repo.

We indeed really need a better way to reuse such code as those dispatching services. It unfortunately seems it would require quite some refactoring throughout the Wikibase. So sadly, although obviously worth it, it would take significant amount of time to do it right.
Thus the really not very clean approach with the optional dependency was proposed. The way it is introduced does not seem that hard to be removed so I hope it would not add on top of the work that should be done during the refactoring.

I've just noticed that I also didn't link the phabricator ticket in the commit message. The ticket might be a better place for the discussion (although I don't say we should not discuss here - simply this patch on Gerrit might be just too "small" for such a general discussion).

Patch-set: 15
Label: Code-Review=0