Update patch set 1
Patch Set 1: Code-Review+1

I believe this change is OK. For the same reason the comment in line 127 (line 129 in the "base" version) could be removed as well.

Not +2 yet as it seems with the current interface split cache-related wrappers (this class and the CachingPropertyInfoLookup) have no in-memory cache the old CachingPropertyInfoStore used to have (this $this->propertyInfo is the remainder of it that was accidentally left over). This might be affecting the performance but on the quick look I cannot see how this could be reasonably implemented given we've split the interfaces.
The change here is fine I think, I'd just like to think about this whole thing a bit more before I dare to merge this.

Patch-set: 1
Label: Code-Review=+1