Update patch set 1
Patch Set 1:

I'm not sure what the purpose of this validation is. In general, we just ignore extra bits of JSON. 

If we do want to fail when the input contains unknown keys, perhaps the deserializers could each declare what keys they support/consume, so we can use that for the check.

But my gut feeling is that this check really serves no purpose.

Patch-set: 1