Update patch set 1
Patch Set 1:

The jenkins failure is intersting, the title is "removing valid claim with no guid fails" but the exxpected failure has nothing to do with not having GUID but actually for having 'remove' option in the data. I'm not sure if I need to delete or not.

Patch-set: 1
Label: Verified=0