Update patch set 5
Patch Set 5:

(8 comments)

It took me a while to realize this code is basically extracting Api\EditEntity::validateMultilangArgs, mostly unchanged. This is fine, I believe, not to try to "improve" that code too much at this point.
Actually, in general Api\EditEntity could also start using this class (ie. the private method of the API class could be removed) but as this patch is part of Api\EditEntity this change will happen any way, so I don't think it must necessarily happen in this patch already.

As I am not very familiar with change op (de)serialization I got confused multiple times while looking at this rather small patches. This resulted in rather lengthy comments. Sorry for that.

I've made few remarks but there is nothing that would be essentially blocking this. Mostly documentation issues. It looks like it would be ready to go pretty soon!

Patch-set: 5