Update patch set 1
Patch Set 1:

I would like to understand why an operation as trivial as a purge needs a token? I believe this should be executed as fast as possible. It is even allowed to fail in this specific case, because there is no code executed when this API request finished (or failed).

Patch-set: 1
Reviewer: Thiemo Kreuz (WMDE) <1453@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Verified=0