Replace FingerprintProvider with more specific ones in RDF builders
This code is still problematic because it is unable to support new types
of terms. But this patch solves one problem: Entity implementations do
not have to support Fingerprints any more, but are free to mix and match
Labels, Descriptions and Aliases as they wish.

Note that one line of code moved outside of the condition. Is this
relevant in terms of performance?

Change-Id: I1c7e71d46a7fef09f6ed1334859474b16a25b73d