Replace FingerprintProvider check in GenericEntityInfoBuilder
This does not solve the problem this class (and other code) does have:
It is still not possible to introduce new term types. But this patch
solves one problem: New entity types do not have to implement the
Fingerprint interface any more, but are free to mix and match Labels,
Descriptions and Aliases as they wish.

Change-Id: I57bbf386de60f1ed570100af5b58cb1008039519