Update patch set 3
Patch Set 3:

I have not looked at the patch itself, but I would like to provide a bit of context anyway. 

The purpose of ApiErrorReporter is to do two things:

1) provide a filly expanded and localized HTML error message to the client.
2) provide the message key and parameter list to the client.

This way, they client can choose to either show the localized error message directly, which should provide information about what failed and how it failed in a user readable way.

Or it can choose to provide it's own error message, based on the message key, the parameters, and the knowledge of what it was trying to do when the error occurred. This allows a less technical, more "natural", error message, tailored to the user's intent.

As far as I know, core does not provide a mechanism that would allow us to do these two things, so we need to either change our frontend, or continue to implement our own error reporting mechanism on top of core's.

Is this a fair summary of the situation?

Patch-set: 3