Update patch set 3
Patch Set 3:

That is a summary of how the situation used to be. Now the client can choose fully expanded and localized HTML by specifying errorformat=html (and possibly errorsuselocal=1 if MediaWiki-namespace customizations should be used), or message key and parameters by specifying errorformat=raw.

Or is your requirement that the client chooses HTML versus raw only after receiving the response rather than knowing which it will need beforehand?

Patch-set: 3