Type hint FingerprintSearchTextGenerator against individual providers
This does not solve all issues (it's still not possible to index other
term types that are not labels, descriptions and aliases), but it makes it
possible to freely mix and match labels, descriptions and aliases in new
entity types without the need to implement the fingerprint interface.

Note that MediaInfo will break with this change and must be adopted.
Patch is already uploaded.

Change-Id: I7c79f7c0d78209d56b271aadf437f3c1ce33283e