Update patch set 7
Patch Set 7:

The summary from our discussion on IRC:

mediawiki/core configures the merge plugin to not merge development dependencies ( composer.json has extra.merge-plugin.merge-dev = false

That is because mediawiki/core and extensions can have different version of PHP CodeSniffer.

You are absolutely right that hamcrest should not be added to require since that is only to run Wikibase tests.

There is no good solution to add test dependencies in an extension. Jenkins runs the tests from the root of mediawiki/core.  I don't think we can just "composer update" in each extension directories, I dont see how the resulting autoloader files would be used and there will be bunch of duplicates.

So I dont have a nice way to solve it but a workaround is to add the dev dependencies directly in mediawiki/core. They are to replace PHPUnit depreacted assertTag() (T69122) which mediawiki/core happens to use and thus it would eventually need those dependencies as well.

So potentially:

* craft a new change for mediawiki/core that adds the two dependencies in require-dev and send patch to Gerrit
* Amend this Wikibase change to remove the deps from require-dev
* In the commit message had a special header to make your Wikibase change to depend on the MediaWiki core change crafted above:

  Depends-On:  <Change id of the mediawiki/core change here>

* send the amended Wikibase patch to Gerrit.

The CI system will find the Depends-On header and will apply the MediaWiki core patch AND the Wikibase patch.  The composer update would then install the require-dev from core and the Wikibase tests should pass.

Patch-set: 7