Register ContentHandlers in $wgContentHandlers.
Previously, Wikibase provided content handlers dynamically,
instead of registering them in the $wgContentHandlers global.
This caused these handlers to be unknown to ContentHandler::getContentModels().

Note: this introduces some global state back into EntityContentFactory.
This is needed to guarantee consistency between EntityContentFactory
and the global state maintained in the ContentHandler class.

Bug: T155139
Change-Id: I73ab3f156427abe206ded671c4a9145924a2c479