Introduce ConfigurableTermSearchInteractor interface
The interface is intended for TermSearchInteractors that could
be configured using TermSearchOptions.
The interface is used to type hint against in classes
that require to pass search options to TermSearchInteractor,
instead of type hinting against the particular implementation
(ie. TermIndexSearchInteractor).

TermIndexSearchInteractor implements the new interface with no
changes needed.

Along with this change deprecated public methods of TermIndexSearchInteractor
for setting individual term search options are replaced with
more generic setTermSearchOptions defined by the interface. This
method has already been in TermIndexSearchInteractor but was
not used.

As no longer needed, deprecated setters are removed from
TermIndexSearchInteractor.

DispatchingTermSearchInteractor also implements the new interface,
and passes current search options to the interactor
actually used in the given term search.

Change-Id: I8db25f5b261ce676a78d8b9a481571bf614fc07a