Update patch set 1
Patch Set 1:

I've looked at the code change itself it seems fine. There is a problem with a failing test but this test is no longer valid if we decide to remove those checks.

I don't know whether it is OK to remove those checks in the first place, though. I've tried to figure out why they've been introduced but I rather failed. It dates back to 2013 and I didn't find a clear explanation when and why it was decided.
This validation is only covered by a single test, the failing one. If I get it right the case tested there is when "remove" filed is misplaced (it appears on a wrong level in the JSON array). Interpretation that came to my mind is that statement-related syntax required by wbeditentity is a bit complex so maybe users sometimes make wrong requests, and we have been showing an error instead of just ignoring the "wrong" key - I believe "remove claims" request with misplaced "remove" key is simply a claims edit which doesn't change anything on the entity (ie. a perfectly fine request but noop).
I think if it was just about this claim-specific problem maybe ChangeOpDeserializer for claims/statements could take care of throwing an error in such particular case. It would allow the general whitelist of keys to be removed as suggested here, and in the end allow other entity types also be used in edit entity API (if we decide to keep this validation it would need to become pluggable so Lexeme and other entity types can add their fields). But this all is just a guess, I don't feel confident about those statements at all :)

Patch-set: 1
Label: Code-Review=0