Update patch set 1
Patch Set 1:

Thanks for summarizing Thiemo. I am going to bring this patch up in today's Wikidata Developer Team daily stand up meeting. If there are no objections to Thiemo's change removing those tags here, I am going to hit the blue button.

IRL meeting of the Dev Team is not the perfect forum to discuss such change (excludes people outside of the team, and those team members who are not in the meeting) but I cannot think of better way to move forward with this patch. It has been up for few days, so we're not going to do anything secretly.

And I would very much like to have the crediting and ownership issues discuss in some appropriate place than Gerrit, I agree with Thiemo we should continue to think and talk about it!

Patch-set: 1