Update patch set 3
Patch Set 3:

I'm sorry? I know this is copying existing code. But this is no excuse to copy all mistakes and issues with it. For example I spend a lot of time reducing the code that does not conform to PSR-4. This patch introduces new code that does not conform to PSR-4, I it really does not matter where this was copied from.

Patch-set: 3