Update patch set 4
Patch Set 4:

(1 comment)

I will not touch this patch further. It works, it solves an actual issue, it is tested. Dependency injection can be improved any time later.

Having an interwiki lookup in the constructor of something that is meant to be an "EntityContent factory" is awkward. I don't think it belongs there. The method we are talking about is an "Entity Title lookup", not something that belongs to a "factory". It seems like this entire method does not belong to this class. I would like us to investigate and split this class much more. Currently it does have 16 public methods.

So much for separation of concerns. Tangling the constructor to this would make this harder to refactor.

Patch-set: 4