Update patch set 3
Patch Set 3:

I also wondered why it is not PSR-4 compatible in the first place, but I am not in the position that I can decide for myself, so I tend to emulate existing behavior of the code, to be conform with the overall code base.

Maybe this is an issue of clear communication and standards.
We could spend some days and convert the code base to PSR-4.

Patch-set: 3
Label: Verified=0