Update patch set 4
Patch Set 4:

> I will not touch this patch further. It works, it solves an actual issue, it is 
> tested. Dependency injection can be improved any time later.

This is adding technical debt. Do we have a good reason to go into debt at this point? Are we committed to paying it back? (This is a small debt - but we should *always* ask ourselves this question).


> Having an interwiki lookup in the constructor of something that is meant to be an "EntityContent factory" is awkward. I don't think it belongs there. The method we are talking about is an "Entity Title lookup", not something that belongs to a "factory". It seems like this entire method does not belong to this class. I would like us to investigate and split this class much more. Currently it does have 16 public methods.

Or the class is misnamed, and everything else is fine. How about EntityContentModelMapper?

In any case - *hiding* the dependency because you think it doesn't belong here doesn't make things better, it makes things worse. The dependency doesn't go away, it's just obscure.

> So much for separation of concerns.

Perhaps the class does to much, yes. Though the concern of knowledge locality may counter separation of concerns. Interface segregation is important, but imho implementations should be centered around shared knowledge (in this case, the mapping between entity types, content models, and namespaces).

Implementation of methods should be grouped into classes based on what they depend on (knowledge, configuration, services). I'm totally open to restructuring this class based on this criterion.

> Tangling the constructor to this would make this harder to refactor.

I disagree: explicitly declared dependencies makes this kind of refactoring easier. Also, explicit injection of mocks in tests makes this refactoring easier.

Patch-set: 4