Update patch set 19
Patch Set 19:

So I've talked to Daniel on how to deal with testing here and we've come to conclusion that applying the change op returned by the deserializer, and checking if the output is correct seems the most reasonable way to deal with this.

We considered that semantically change op deserialzer is used to make some particular change to the entity, that's why it seems to make sense to test that the deserializer works by checking if changes it /deserializes/ have an expected effect on the entity.
This means potentially doing similar things in tests of the particular change op, and a change op deserializer but for this we should be able to have some solution to avoid duplicating work or doing any other silly things. For instance deserializer test could be somehow aware of the relevant change op test, or we could store test cases in external files (ie. JSON serializations of the input entity, the change op, and the expected entity after the change). Thing to consider here would be to how much of edge cases should be actually tested in change op deserializer test.

One more argument in favour of this approach would be that Change op deserializer tests wouldn't need to be changed if we ever decided to change ChangeOp classes (e.g. split them, merge them etc). On the contrary those tests would assure the changes in the other layer (ie. change ops) doesn't break anything.

Patch-set: 19
Label: Verified=0